Temporal fix for memory problem on models.modelBase.AlignmentModel.keyDiv
This bit is very nasty as Numpy takes far to much memory for this
function. Right now this solution would be enough on macOS because
macOS is equipped with very impressive memory compression and
virtualisation which makes this issue sometimes even unnoticeable, but
for Linux, well they are going to seek other solutions.